Use Float64

+++ experiments/AtmosGCM/solidbody.jl
@@ -94,7 +94,7 @@ end
 
 function main()
     # Driver configuration parameters
-    FT = Float32                             # floating type precision
+    FT = Float64                             # floating type precision
     poly_order = 5                           # discontinuous Galerkin polynomial order
     n_horz = 3                               # horizontal element number
     n_vert = 3                               # vertical element number
